Fix occasional crash of Sublime Text
Sublime Text can crash on certain layout changes, especially if
another layout change is still on the fly.

 see:  https://github.com/SublimeTextIssues/Core/issues/1785

This commit queues/delays the execution of the code in the
`on_activated()` handler to ensure the ST core can finish a layout
change before calling the next one.

A synchronous `set_timeout()` is used for backward compatibility with
ST2 and because of no flickering when manipulating the layout. The
10ms delay is also added for backward compatibility reasons and is not
needed for recent ST3 builds.

The `view` is passed using the `partial()` function to ensure the
queued worker uses the correct and valid instance.

Note:

  The event handlers don't need to return a value like `None`.

Fix partial doesn't work

The on_activated() is not called if partial() is used in the decorator.
Hence, remove it.

Fix on_window_command

Queue `on_window_command` the same way as `on_activate` causes strange
things to happen. Revert it.

Simplify decorator

The decorator's only job is to run a function delayed.