Update README to announce focus_neighboring_group
Sublime Text 3 provides the `focus_neighboring_group` command out of
the box, which replaces `shift_pane` and `unshift_pane`.

The implementation is nearly the former one uses the boolean argument
`forward` to decide whether to shift to the previous or next pane.

The README is updated to encourage users to switch to the
`focus_neighboring_group` command.