Add a workaround for core issue 2919
see: https://github.com/SublimeTextIssues/Core/issues/2919

Sublime Text does not correctly set the focus to a transient view, if
it is the only one in a group. As a result, navigating to such a group
is not possible, if another one is maximized.

The `set_layout_and_focus()` sets focus to the active group after
changing the layout manually to work around this bug.

Credits:
  Thanks to @evandrocoan who found the bug and created this workaround.